Fix two trade-visibility gaps: orphans never permanently recorded, real_position_ticket 32-bit overflow
Prompted by the user's explicit requirement that no placed trade or
open position ever be missing from the UI, per user/model/account.

Bug 1: an orphaned real position that got self-healed (target
attached) was never written to the trades table -- only a successful
close-detection later would create a row. If the position closed
before that ever ran (exactly what happened to the two positions
found this session), it vanished from trade history permanently, with
no record it ever existed. Fixed by writing a permanent Trade row
(write_orphan_trade(), real_status='open') the moment an orphan is
found, independent of whether healing itself succeeds, and handing it
off to ongoing tracking so its eventual close lands on the same row.

Bug 2 (larger, found while building the above): trades.real_position_
ticket was a plain 32-bit Postgres Integer (max 2,147,483,647). Every
real MT5 ticket observed this session already exceeds that. Confirmed
live: zero rows in production have ever had this column populated,
despite weeks of real trading -- consistent with every real trade
write silently failing at the DB level (discarding the whole row, not
just this field, since it's one INSERT) since the broker's ticket
numbers crossed 2^31. Migration 0022 widens it to BigInteger; audited
every other Integer column in the schema to confirm this is the only
one holding broker-assigned (not app-chosen) numbers.

check_for_orphaned_positions() now takes risk_pct and returns
trade_id/entry_time_ny per result; _heal_orphan() now returns the
computed target even when the broker-side modify fails, so a record
can still be written with a real target instead of falling back to
open_price. New event type orphan_trade_recorded (distinct from
orphan_position_recovered -- finding/healing and recording are two
independent protections, either can fail without disabling the
other).

Full suite: 405 passed, 1 skipped, 10 pre-existing unrelated failures
(same baseline as before this change) -- zero regressions.

+++ app/models/trade.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Boolean, CheckConstraint, Column, DateTime, Float, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Boolean, CheckConstraint, Column, DateTime, Float, ForeignKey, String
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import relationship
 
@@ -71,7 +71,13 @@ class Trade(Base):
     # only populate once OrderManager's real fill/close detection has
     # actually happened (see shadow_runner/order_manager.py's
     # get_real_outcome()).
-    real_position_ticket = Column(Integer, nullable=True)
+    # BigInteger since migration 0022 -- was a plain 32-bit Integer
+    # (max 2,147,483,647), which every real MT5 ticket actually observed
+    # 2026-09-04 already exceeded, silently failing every real trade
+    # write that reached this column (discarding the whole row, not just
+    # this field, since it's one INSERT) -- confirmed live: zero rows
+    # ever had this populated, despite weeks of real trading.
+    real_position_ticket = Column(BigInteger, nullable=True)
     real_fill_price = Column(Float, nullable=True)
     real_fill_time_utc = Column(DateTime(timezone=True), nullable=True)
     real_fill_time_ny = Column(DateTime(timezone=True), nullable=True)
